+++ src/test/java/spoon/test/intercession/IntercessionContractTest.java
@@ -5,21 +5,12 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import spoon.Launcher;
 import spoon.processing.FactoryAccessor;
-import spoon.reflect.ast.IntercessionScanner;
-import spoon.reflect.declaration.CtClass;
 import spoon.reflect.declaration.CtMethod;
-import spoon.reflect.declaration.CtType;
-import spoon.reflect.declaration.CtTypedElement;
-import spoon.reflect.declaration.ModifierKind;
 import spoon.reflect.factory.Factory;
 import spoon.reflect.factory.FactoryImpl;
-import spoon.reflect.reference.CtTypeReference;
-import spoon.reflect.visitor.Query;
-import spoon.reflect.visitor.filter.TypeFilter;
 import spoon.support.DefaultCoreFactory;
 import spoon.support.StandardEnvironment;
 
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Collection;
